lnwallet: add test for spending from commitment txn, tests Bitcoin Core's BIP 68 implementation
An active Bitcoin Core node running in regtest mode is required for
this test to pass.

In order to initialize the state of the node for the test, run the
following commands:
  * bitcoin-cli -regtest generate 101
  * bitcoin-cli -regtest sendtoaddress
"mrsU9wTxs1UB4wqWZbuA6Nd6uYxs2VKe8T" 12
  * bitcoin-cli -regtest getrawtransaction <txid ret from above here>

With the txid of the funds sent to alice in hand, fill in lines 74-84
w/ the txid and output number for the output created for alice.

After all the set up has been completed. Simply run the test with:
  * go test -run=TestCSVCommitmentTransaction