can close multichans and remove from db
but actually we can't delete from the db, we have to flag and store
the spending tx, etc.  Also there are problems with multiple concurrent
multixos.  It's probably sorting of the multixos (I don't think there
IS any sorting yet... so whatever boltdb produces, which will probably
be different for each computer)

So fix that stuff next.