further simplify elkrem, revise pushpull, stateTX
Elkrem sender doesn't need state; it's not even really
a struct, it's just a single root hash, and the only thing
it can do its descend down to a leaf.  So there's no need
for serialization, it's just 32 bytes.  Which isn't saved
anyhere anyway, it's derived.

This may break some stuff in lnwallet but it can probably
be replaces with just writing 32 bytes and restoring from that.

Also... push / pull sequence...