start closeTxo storage.
It's actually tricky because there can be 2 spendable outputs
from 1 channel (if they broadcast an old state).  Or 1
spendable output.  Or 0.

Getting there but will do it tomorrow.

Also I've been saving the channel outpoints in the bucket
which is a waste of space so should get rid of that.