Use pubkeys instead of pubkey hashes for refund
Next, have the close txs use the refund pubkey for
CSV instead of the channel pubkey.  This makes it
so that when a break tx occurs, nobody can tell who
broke it (other than the channel parties)

HAKD pubkey should still be based on the channel pubkey.
If it's unrevoked, nobody will know what it's based on
(not even the potential grabber).  If it can be grabbed though,
the grabber can prove to everyone the invalid state by publishing
the elkrem hash.