fixed some HAKD bugs, it seems to work now, but I have to redo everything...
I'm pretty sure adding the elkrem point doesn't work to hide the channel
from the sorceror.  The sorceror can take the pubkeys from the commit
script and subtract the base points they know about; if the results from
the timeout and revoke operation are the same, they know that was the channel
they were watching.
I think there has to be 2 elkrems... argh.